# docs/swagger.yaml
basePath: /
definitions:
  models.Item:
    description: Shopping item information
    properties:
      id:
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
      name:
        example: Milk
        type: string
    type: object
  models.User:
    properties:
      created_utc:
        type: string
      deleted_utc:
        type: string
      email:
        type: string
      id:
        type: string
      profile_picture:
        type: string
      user_name:
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: This is the base service for boondocks API
  title: Boondock-Service-Base API
  version: "1.0"
paths:
  /items:
    get:
      description: Get full shopping list
      produces:
      - application/json
      responses:
        "200":
          description: Returns the shopping list, empty array if no items
          schema:
            items:
              $ref: '#/definitions/models.Item'
            type: array
      summary: List all shopping items
      tags:
      - items
  /items/:
    post:
      description: Creates a new shopping item from provided json data
      parameters:
      - description: Item
        format: models.Item
        in: body
        name: name
        required: true
        schema:
          $ref: '#/definitions/models.Item'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: ""
      summary: Create a new shopping item
      tags:
      - items
  /items/{id}:
    delete:
      description: Removes a shopping item from storage by UUID
      parameters:
      - description: Item ID
        example: 550e8400-e29b-41d4-a716-446655440000
        format: uuid
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Item found
          schema:
            $ref: '#/definitions/models.Item'
        "400":
          description: Invalid UUID format
          schema:
            type: string
        "404":
          description: Item not found
          schema:
            type: string
      summary: Deltete a shopping item by id
      tags:
      - items
    get:
      description: Gets a shopping item by UUID
      parameters:
      - description: Item ID
        example: 550e8400-e29b-41d4-a716-446655440000
        format: uuid
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Item found
          schema:
            $ref: '#/definitions/models.Item'
        "400":
          description: Invalid UUID format
          schema:
            type: string
        "404":
          description: Item not found
          schema:
            type: string
      summary: Get a shopping item by id
      tags:
      - items
  /users:
    get:
      description: Get all users in the database
      produces:
      - application/json
      responses:
        "200":
          description: Returns empty array if no users found
          schema:
            items:
              $ref: '#/definitions/models.User'
            type: array
      summary: List all users
      tags:
      - users
  /users/{id}:
    get:
      description: Gets a user by UUID
      parameters:
      - description: User ID
        example: 550e8400-e29b-41d4-a716-446655440000
        format: uuid
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: User found
          schema:
            $ref: '#/definitions/models.User'
        "400":
          description: Invalid UUID format
          schema:
            type: string
        "404":
          description: User not found
          schema:
            type: string
      summary: Get a user by id
      tags:
      - users
swagger: "2.0"